+++ tests/phpunit/src/Commands/Push/PushCodeCommandTest.php
@@ -7,7 +7,7 @@ use Acquia\Cli\Command\Pull\PullCodeCommand;
 use Acquia\Cli\Command\Pull\PullCommand;
 use Acquia\Cli\Command\Pull\PullDatabaseCommand;
 use Acquia\Cli\Command\Pull\PullFilesCommand;
-use Acquia\Cli\Command\Push\PushCodeCommand;
+use Acquia\Cli\Command\Push\PushArtifactCommand;
 use Acquia\Cli\Command\Push\PushFilesCommand;
 use Acquia\Cli\Exception\AcquiaCliException;
 use Acquia\Cli\Helpers\SshHelper;
@@ -24,7 +24,7 @@ use Webmozart\PathUtil\Path;
 /**
  * Class PushCodeCommandTest.
  *
- * @property \Acquia\Cli\Command\Push\PushCodeCommand $command
+ * @property \Acquia\Cli\Command\Push\PushArtifactCommand $command
  * @package Acquia\Cli\Tests\Commands\Push
  */
 class PushCodeCommandTest extends CommandTestBase {
@@ -33,7 +33,7 @@ class PushCodeCommandTest extends CommandTestBase {
    * {@inheritdoc}
    */
   protected function createCommand(): Command {
-    return $this->injectCommand(PushCodeCommand::class);
+    return $this->injectCommand(PushArtifactCommand::class);
   }
 
   public function testPushCode(): void {
